+++ pkg/analysis_server/lib/src/cider/local_library_contributor.dart
@@ -61,14 +61,21 @@ class LibraryElementSuggestionBuilder
   @override
   void visitClassElement(ClassElement element) {
     AstNode node = request.target.containingNode;
-    var libraryElement = request.libraryElement;
-    if (node is ExtendsClause && !element.isExtendableIn(libraryElement)) {
-      return;
-    } else if (node is ImplementsClause &&
-        !element.isImplementableIn(libraryElement)) {
-      return;
-    } else if (node is WithClause && !element.isMixableIn(libraryElement)) {
-      return;
+    if (node is ExtendsClause) {
+      if (element.library != request.libraryElement &&
+          !element.isExtendableOutside) {
+        return;
+      }
+    } else if (node is ImplementsClause) {
+      if (element.library != request.libraryElement &&
+          !element.isImplementableOutside) {
+        return;
+      }
+    } else if (node is WithClause) {
+      if (element.library != request.libraryElement &&
+          !element.isMixableOutside) {
+        return;
+      }
     }
     _visitInterfaceElement(element);
   }
@@ -128,7 +135,8 @@ class LibraryElementSuggestionBuilder
   visitMixinElement(MixinElement element) {
     AstNode node = request.target.containingNode;
     if (node is ImplementsClause &&
-        !element.isImplementableIn(request.libraryElement)) {
+        request.libraryElement != element.library &&
+        !element.isImplementableOutside) {
       return;
     }
     _visitInterfaceElement(element);